requirements.txt
openai==0.27.2
python-telegram-bot==20.1
pydub==0.25.1
python-dotenv